Problem 3 "Largest prime factor": solved

+++ problem0003/src/main/java/problem0003/App.java
@@ -4,7 +4,7 @@ public class App {
 
   public static void main(String[] args) {
     final LargestPrimeFactor calculator = new LargestPrimeFactor();
-    final int sum = calculator.compute(600851475143L);
+    final long sum = calculator.compute(600851475143L);
 
     System.out.println(sum);
   }
